+++ babel.config.js
@@ -5,7 +5,7 @@ module.exports = api => {
     [
       '@babel/preset-env',
       {
-	// https://browserl.ist/?q=last+4+Chrome+version%2C+last+2+Firefox+version
+        // https://browserl.ist/?q=last+4+Chrome+version%2C+last+2+Firefox+version
         targets: 'last 4 Chrome version, last 2 Firefox version',
         useBuiltIns: 'entry',
         corejs: 2,
@@ -16,6 +16,8 @@ module.exports = api => {
   ]
   const plugins = [
     '@babel/plugin-proposal-class-properties',
+    '@babel/plugin-proposal-optional-chaining',
+    '@babel/plugin-proposal-nullish-coalescing-operator',
   ]
 
   return {